\name{outlieranalysis}
\alias{outlieranalysis}
\title{outlieranalysis}
\usage{
Useful function for missing value and outlier analysis.
}
\description{
This functions accepts dataframe and returns percentage missing values in each column. It also indicates number of outliers present in each column.
}
\examples{
outlieranalysis(data)
}
